refactor: Calculator 클래스 -> Calculator, Operator 클래스로 분리
