+++ src/test/kotlin/calculator/model/CalculatorTest.kt
@@ -11,24 +11,6 @@ import org.junit.jupiter.params.provider.CsvSource
 class CalculatorTest {
 
     @Test
-    fun `존재하지 않는 수식 기호 예외 발생`() {
-        // given, when, then
-        val exception = assertThrows(IllegalArgumentException::class.java) {
-            Calculator.calculate("1 & 2")
-        }
-        assertEquals("존재하지 않는 수식 기호입니다. (symbol: &)", exception.message)
-    }
-
-    @Test
-    fun `나누는 수가 0일 때 예외 발생`() {
-        // given, when, then
-        val exception = assertThrows(IllegalArgumentException::class.java) {
-            Calculator.calculate("1 / 0")
-        }
-        assertEquals("나누는 수가 0이 될 수 없습니다.", exception.message)
-    }
-
-    @Test
     fun `수식이 공백일 때 예외 발생`() {
         // given, when, then
         val exception = assertThrows(IllegalArgumentException::class.java) {
